Extract base residual wrappers
