+++ applications/sintering/include/pf-applications/sintering/residual_wrapper.h
@@ -15,7 +15,7 @@
 
 #pragma once
 
-#include <pf-applications/lac/solvers_nonlinear.h>
+#include <pf-applications/lac/residual_wrapper.h>
 
 #include <pf-applications/sintering/operator_advection.h>
 
@@ -25,108 +25,6 @@ namespace Sintering
 {
   using namespace dealii;
   using namespace NonLinearSolvers;
-
-  template <typename Number>
-  class ResidualWrapper : public Subscriptor
-  {
-  public:
-    using value_type  = Number;
-    using vector_type = LinearAlgebra::distributed::Vector<Number>;
-    using VectorType  = vector_type;
-    using BlockVectorType =
-      LinearAlgebra::distributed::DynamicBlockVector<Number>;
-
-    // This mimic the operator's interface
-    template <int with_time_derivative = 2>
-    void
-    evaluate_nonlinear_residual(
-      BlockVectorType       &dst,
-      const BlockVectorType &src,
-      const std::function<void(const unsigned int, const unsigned int)>
-        pre_operation = {},
-      const std::function<void(const unsigned int, const unsigned int)>
-        post_operation = {}) const
-    {
-      static_assert(with_time_derivative == 1 || with_time_derivative == 2);
-
-      if constexpr (with_time_derivative == 2)
-        evaluate_for_residual(dst, src, pre_operation, post_operation);
-      else if (with_time_derivative == 1)
-        evaluate_for_jacobian(dst, src, pre_operation, post_operation);
-    }
-
-  private:
-    virtual void
-    evaluate_for_residual(
-      BlockVectorType       &dst,
-      const BlockVectorType &src,
-      const std::function<void(const unsigned int, const unsigned int)>
-        pre_operation = {},
-      const std::function<void(const unsigned int, const unsigned int)>
-        post_operation = {}) const = 0;
-
-    virtual void
-    evaluate_for_jacobian(
-      BlockVectorType       &dst,
-      const BlockVectorType &src,
-      const std::function<void(const unsigned int, const unsigned int)>
-        pre_operation = {},
-      const std::function<void(const unsigned int, const unsigned int)>
-        post_operation = {}) const = 0;
-  };
-
-  template <typename Number, typename OperatorType>
-  class ResidualWrapperBase : public ResidualWrapper<Number>
-  {
-  public:
-    using value_type      = typename ResidualWrapper<Number>::value_type;
-    using vector_type     = typename ResidualWrapper<Number>::vector_type;
-    using VectorType      = typename ResidualWrapper<Number>::VectorType;
-    using BlockVectorType = typename ResidualWrapper<Number>::BlockVectorType;
-
-    ResidualWrapperBase(const OperatorType &nonlinear_operator)
-      : nonlinear_operator(nonlinear_operator)
-    {}
-
-  protected:
-    template <unsigned int with_time_derivative = 2>
-    void
-    do_evaluate_nonlinear_residual(
-      BlockVectorType       &dst,
-      const BlockVectorType &src,
-      const std::function<void(const unsigned int, const unsigned int)>
-        pre_operation = {},
-      const std::function<void(const unsigned int, const unsigned int)>
-        post_operation = {}) const
-    {
-      if constexpr (has_evaluate_nonlinear_residual_with_pre_post<
-                      OperatorType,
-                      BlockVectorType>)
-        {
-          nonlinear_operator
-            .template evaluate_nonlinear_residual<with_time_derivative>(
-              dst, src, pre_operation, post_operation);
-        }
-      else
-        {
-          std::pair<unsigned int, unsigned int> range{
-            0, nonlinear_operator.get_matrix_free().n_cell_batches()};
-
-          if (pre_operation)
-            pre_operation(range.first, range.second);
-
-          nonlinear_operator
-            .template evaluate_nonlinear_residual<with_time_derivative>(dst,
-                                                                        src);
-
-          if (post_operation)
-            post_operation(range.first, range.second);
-        }
-    }
-
-  private:
-    const OperatorType &nonlinear_operator;
-  };
 
   template <typename Number, typename OperatorType>
   class ResidualWrapperGeneric
